+++ packages/orq-rc/src/funcs/knowledgeCreateChunks.ts
@@ -34,7 +34,7 @@ export function knowledgeCreateChunks(
   options?: RequestOptions,
 ): APIPromise<
   Result<
-    Array<operations.CreateChunkResponseBody>,
+    Array<operations.ResponseBody>,
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -59,7 +59,7 @@ async function $do(
 ): Promise<
   [
     Result<
-      Array<operations.CreateChunkResponseBody>,
+      Array<operations.ResponseBody>,
       | OrqError
       | ResponseValidationError
       | ConnectionError
@@ -148,7 +148,7 @@ async function $do(
   const response = doResult.value;
 
   const [result] = await M.match<
-    Array<operations.CreateChunkResponseBody>,
+    Array<operations.ResponseBody>,
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -158,7 +158,7 @@ async function $do(
     | UnexpectedClientError
     | SDKValidationError
   >(
-    M.json(200, z.array(operations.CreateChunkResponseBody$inboundSchema)),
+    M.json(200, z.array(operations.ResponseBody$inboundSchema)),
     M.fail("4XX"),
     M.fail([500, "5XX"]),
   )(response, req);