+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-14T13:10:49.498Z",
+    "2026-04-15T05:48:02.893Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {