+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -403,7 +403,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KP61VV20VRB4TH7HYES99PHN"),
+  _id: z.string().default("01KP7TXVA4TZW046KF0GBN56VX"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,